devastating changes applied, gaaah

+++ superadmin/delete_internship.php
@@ -5,7 +5,14 @@ require_once '../includes/functions.php';
 
 requireAdminAreaAccess();
 
-$id = $_GET['id'] ?? 0;
+if ($_SERVER['REQUEST_METHOD'] !== 'POST') {
+    header('Location: manage_internships.php');
+    exit;
+}
+
+requireValidCsrfToken(['redirect' => 'manage_internships.php']);
+
+$id = isset($_POST['id']) ? (int) $_POST['id'] : 0;
 
 $stmt = $pdo->prepare("DELETE FROM applications WHERE internship_id = ?");
 $stmt->execute([$id]);
